+++ Integrations/Panorama/Panorama.py
@@ -2067,21 +2067,28 @@ def panorama_edit_custom_url_category_command():
 
 
 @logger
-def panorama_get_url_category(url):
+def panorama_get_url_category(url_cmd, url):
     params = {
         'action': 'show',
         'type': 'op',
         'key': API_KEY,
-        'cmd': '<test><url>' + url + '</url></test>'
+        'cmd': f'<test><{url_cmd}>{url}</{url_cmd}></test>'
     }
-    result = http_request(
+    raw_result = http_request(
         URL,
         'POST',
         params=params,
     )
-
-    s = result['response']['result'].splitlines()[1]
-    return s.split(' ')[1]
+    result = raw_result['response']['result']
+    if url_cmd == 'url-info-host':
+        category = result.split(': ')[1]
+    else:
+        result = result.splitlines()[1]
+        if url_cmd == 'url':
+            category = result.split(' ')[1]
+        else:  # url-info-cloud
+            category = result.split(',')[3]
+    return category
 
 
 def populate_url_filter_category_from_context(category):
@@ -2099,7 +2106,7 @@ def populate_url_filter_category_from_context(category):
             return context_urls
 
 
-def panorama_get_url_category_command():
+def panorama_get_url_category_command(url_cmd: str):
     """
     Get the url category from Palo Alto URL Filtering
     """
@@ -2107,7 +2114,7 @@ def panorama_get_url_category_command():
 
     categories_dict: Dict[str, list] = {}
     for url in urls:
-        category = panorama_get_url_category(url)
+        category = panorama_get_url_category(url_cmd, url)
         if category in categories_dict:
             categories_dict[category].append(url)
         else:
@@ -2121,17 +2128,25 @@ def panorama_get_url_category_command():
             'Category': key,
             'URL': value
         })
+    title = 'URL Filtering'
+    if url_cmd == 'url-info-cloud':
+        title += f' from cloud'
+    elif url_cmd == 'url-info-host':
+        title += f' from host'
+    human_readable = tableToMarkdown(f'{title}:', url_category_output, ['URL', 'Category'], removeNull=True)
 
     demisto.results({
         'Type': entryTypes['note'],
         'ContentsFormat': formats['json'],
         'Contents': categories_dict,
         'ReadableContentsFormat': formats['markdown'],
-        'HumanReadable': tableToMarkdown('URL Filtering:', url_category_output, ['URL', 'Category'], removeNull=True),
+        'HumanReadable': human_readable,
         'EntryContext': {
             "Panorama.URLFilter(val.Category === obj.Category)": url_category_output
         }
     })
+
+''' URL Filter '''
 
 
 def prettify_get_url_filter(url_filter):
@@ -4217,8 +4232,15 @@ def main():
 
         # URL Filtering capabilities
         elif demisto.command() == 'panorama-get-url-category':
-            panorama_get_url_category_command()
+            panorama_get_url_category_command(url_cmd='url')
 
+        elif demisto.command() == 'panorama-get-url-category-from-cloud':
+            panorama_get_url_category_command(url_cmd='url-info-cloud')
+
+        elif demisto.command() == 'panorama-get-url-category-from-host':
+            panorama_get_url_category_command(url_cmd='url-info-host')
+
+        # URL Filter
         elif demisto.command() == 'panorama-get-url-filter':
             panorama_get_url_filter_command()